requirements.txt
numpy==1.21.6
opencv-python==4.6.0.66
packaging==22.0
Pillow==9.4.0
pip==21.3.1
pytesseract==0.3.10
setuptools==60.2.0
wheel==0.37.1

imutils~=0.5.4
Flask~=2.2.2